PORtfolio SECtion and  resolve issue  with deleteing categories

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,10 +3,11 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Models\Category;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use App\Models\PortfolioItem;
 use App\Http\Controllers\Controller;
 use App\DataTables\CategoryDataTable;
-use Illuminate\Support\Str;
 
 class CategoryController extends Controller
 {
@@ -92,6 +93,13 @@ class CategoryController extends Controller
     {
         //
         $category = Category::findOrFail($id);
-        $category->delete();
+        $hasItem = PortfolioItem::where('category_id', $category->id)->count();
+        if($hasItem === 0){
+            $category->delete();
+            return true;
+        }
+        
+        return response(['status' => 'error']);
+        // $category->delete();
     }
 }
